requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.9
charset-normalizer==3.4.2
cloudinary==1.44.1
dj-database-url==3.0.1
Django==5.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-environ==0.11.2
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.27.1
drf-yasg==1.21.7
ecdsa==0.19.1
environ==1.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-http-client==3.3.7
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
sendgrid==6.12.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
channels==4.0.0
channels-redis==4.1.0
cryptography==41.0.0
redis==5.0.1
django-ratelimit==3.0.1
django-redis==5.2.0
python-socketio==5.9.0
eventlet==0.34.3
gunicorn==21.2.0
daphne==4.0.0
uvicorn==0.27.0
